16_редактирет поведение свайпа списка
