+++ app/src/main/java/com/bryukhanov/shoppinglist/mylists/presentation/view/MyListsFragment.kt
@@ -1,6 +1,5 @@
 package com.bryukhanov.shoppinglist.mylists.presentation.view
 
-import android.graphics.Canvas
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,7 +8,6 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.bryukhanov.shoppinglist.R
 import com.bryukhanov.shoppinglist.databinding.FragmentMyListsBinding
 import com.bryukhanov.shoppinglist.mylists.domain.models.ShoppingListItem
@@ -56,56 +54,15 @@ class MyListsFragment : Fragment() {
             }
         })
 
+        adapter.notifyDataSetChanged()
+
         binding.rvMyLists.apply {
             layoutManager = LinearLayoutManager(context)
             setHasFixedSize(true)
             adapter = this@MyListsFragment.adapter
         }
 
-        val itemTouchHelper = ItemTouchHelper(object : ItemTouchHelper.Callback() {
-            override fun getMovementFlags(
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder
-            ): Int {
-                return makeMovementFlags(0, ItemTouchHelper.LEFT)
-            }
-
-            override fun onMove(
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                target: RecyclerView.ViewHolder
-            ): Boolean = false
-
-            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                val position = viewHolder.bindingAdapterPosition
-                adapter.showActions(position)
-            }
-
-            override fun onChildDraw(
-                c: Canvas,
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                dX: Float,
-                dY: Float,
-                actionState: Int,
-                isCurrentlyActive: Boolean
-            ) {
-                val itemView = viewHolder.itemView
-                val buttonContainerWidth = itemView.findViewById<View>(R.id.buttonContainer).width
-                val maxDx = -buttonContainerWidth.toFloat() // Ограничиваем свайп до ширины кнопок
-
-                super.onChildDraw(
-                    c,
-                    recyclerView,
-                    viewHolder,
-                    dX.coerceAtLeast(maxDx), // Свайп ограничен до кнопок
-                    dY,
-                    actionState,
-                    isCurrentlyActive
-                )
-            }
-        })
-
+        val itemTouchHelper = ItemTouchHelper(SwipeCallback(adapter))
         itemTouchHelper.attachToRecyclerView(binding.rvMyLists)
     }
 
@@ -114,4 +71,5 @@ class MyListsFragment : Fragment() {
         _binding = null
     }
 }
+
