+++ app/src/main/res/layout/item_my_list.xml
@@ -22,8 +22,7 @@
             android:layout_height="@dimen/twelve_space"
             android:background="@android:color/transparent"
             android:contentDescription="@string/edit"
-            android:src="@drawable/ic_edit_list"
-            android:visibility="visible" />
+            android:src="@drawable/ic_edit_list" />
 
         <ImageButton
             android:id="@+id/btnCopy"
@@ -31,8 +30,7 @@
             android:layout_height="@dimen/twelve_space"
             android:background="@android:color/transparent"
             android:contentDescription="@string/copy"
-            android:src="@drawable/ic_copy_list"
-            android:visibility="visible" />
+            android:src="@drawable/ic_copy_list" />
 
         <ImageButton
             android:id="@+id/btnDelete"
@@ -41,8 +39,7 @@
             android:layout_gravity="end"
             android:background="@android:color/transparent"
             android:contentDescription="@string/delete"
-            android:src="@drawable/ic_delete_list"
-            android:visibility="visible" />
+            android:src="@drawable/ic_delete_list" />
     </LinearLayout>
 
     <androidx.cardview.widget.CardView
