+++ app/src/main/res/values/dimens.xml
@@ -7,6 +7,7 @@
     <dimen name="four_space">16dp</dimen>
     <dimen name="four_and_half_space">18dp</dimen>
     <dimen name="six_space">24dp</dimen>
+    <dimen name="eight_space">32dp</dimen>
     <dimen name="nine_space">36dp</dimen>
     <dimen name="ten_space">40dp</dimen>
     <dimen name="eleven_space">44dp</dimen>
@@ -18,6 +19,6 @@
     <dimen name="sixteen_size_text">16sp</dimen>
     <dimen name="twenty_two_size_text">22sp</dimen>
     <dimen name="twenty_four_size_text">24sp</dimen>
-    <dimen name="eight_space">32dp</dimen>
+
 
 </resources>
